Refactors the  Game logic into game component and extracts out the header and footer into components
