+++ src/components/Game.vue
@@ -2,20 +2,20 @@
     <v-alert color="accent">
         <v-row>
         <v-col>
-          <h4 class="gameText">Task: {{ text }}</h4>
+          <h4 class="gameText">Task: {{ gameTask.text }}</h4>
         </v-col>
         <v-col>
-        <h4 class="gameText">Score: {{ score }}</h4>
+        <h4 class="gameText">Score: {{ gameScore }}</h4>
         </v-col>
         <v-col>
         <h4 class="gameText">Time Left: {{ secondsLeft }} seconds</h4>
         </v-col>
         </v-row>
       <v-toolbar color="secondary" v-if="gameOver">
-      <v-toolbar-title class="gameOverText">Game Over</v-toolbar-title>
+      <v-toolbar-title class="white--text">Game Over</v-toolbar-title>
       <v-spacer></v-spacer>
       <v-toolbar-items>
-      <v-btn color="secondary" @click="closeGame">Close</v-btn>
+      <v-btn color="secondary" @click="closeGame()">Close</v-btn>
       <v-btn color="secondary" @click="restartGame">Restart</v-btn>
       </v-toolbar-items>
       </v-toolbar>
@@ -23,17 +23,56 @@
 </template>
 
 <script>
+import {
+  getRandomCardMark,
+  getRandomCardNumber,
+  getRandomNumber,
+  coinFlip
+} from '../utils'
+import { marks, gameSeconds } from '../const'
 import { eventBus } from '../main'
 export default {
-  props: [ 'text', 'score' ],
+  created () {
+    this.makeCards()
+    eventBus.$on('makeCards', () => {
+      this.makeCards()
+      eventBus.$emit('MakeCardsReceived')
+    })
+    eventBus.$on('cardWasClicked', (data) => {
+      if (this.gameOn) {
+        this.checkCard(data)
+      }
+    })
+    eventBus.$on('startGameClicked', () => {
+      this.startGame()
+    })
+  },
+  props: ['numberOfCards'],
   data () {
     return {
-      secondsLeft: 30,
+      gameScore: 0,
+      secondsLeft: gameSeconds,
       interval: setInterval(this.timer, 1000),
-      gameOver: false
+      gameOver: false,
+      taskNumber: 0,
+      isMarkTask: false,
+      gameOn: false,
+      cards: []
     }
   },
   methods: {
+    makeCards () {
+      this.cards = []
+      for (let index = 0; index < this.numberOfCards; index++) {
+        const randomMark = getRandomCardMark()
+        const randomNumber = getRandomCardNumber()
+        this.cards.push({
+          mark: randomMark,
+          number: randomNumber
+        })
+      }
+      eventBus.$emit('updateCards', this.cards)
+    },
     timer () {
       this.secondsLeft -= 1
       if (this.secondsLeft < 1) {
@@ -46,13 +85,58 @@ export default {
       clearInterval(this.interval)
     },
     restartGame () {
-      eventBus.$emit('restartGame')
-      this.secondsLeft = 30
+      this.gameScore = 0
+      this.startGame()
+    },
+    startGame () {
       this.gameOver = false
+      this.secondsLeft = gameSeconds
+      this.gameOn = true
+      this.taskNumber = this.getTaskNumber()
+      this.stopTimer()
       this.interval = setInterval(this.timer, 1000)
+      eventBus.$emit('startGame')
     },
     closeGame () {
-      eventBus.$emit('closeGame')
+      eventBus.$emit('closeGameClicked')
+      this.gameScore = 0
+      this.gameOn = false
+    },
+    checkCard (data) {
+      if (this.gameTask.task(data)) {
+        this.gameScore += 100
+      }
+      this.taskNumber = this.getTaskNumber()
+    },
+    getTaskNumber () {
+      this.makeCards()
+      this.isMarkTask = coinFlip()
+      return getRandomNumber(0, this.cards.length - 1)
+    }
+  },
+  computed: {
+    gameTask () {
+      if (this.cards === []) {
+        this.makeCards()
+      }
+      const randomCard = this.cards[this.taskNumber]
+      const markText = marks[randomCard.mark]
+      if (this.isMarkTask) {
+        return {
+          text: 'Find a ' + markText,
+          task (card) {
+            return card.mark === randomCard.mark
+          }
+        }
+      }
+      const numText =
+        randomCard.number === 1 ? 'n Ace' : ' ' + randomCard.number
+      return {
+        text: 'Find a' + numText,
+        task (card) {
+          return card.number === randomCard.number
+        }
+      }
     }
   }
 }
@@ -63,10 +147,6 @@ export default {
   border-radius: 25px;
   background-color:#333;
   padding:10px;
-  color:#fff;
-}
-
-.gameOverText {
   color:#fff;
 }
 </style>
