+++ src/const.js
@@ -3,3 +3,4 @@ export const title = 'Beautiful Cat Facts'
 export const copyright = 'Created by Zach Sarette in 2019'
 export const catFactsUrl = 'https://catfact.ninja/facts?limit='
 export const catPicsUrl = 'https://api.thecatapi.com/v1/images/search?limit='
+export const gameSeconds = 30
